Internal refactor to move defintion of modifierConditions closer to where they're used
