+++ .changeset/spicy-clouds-invite/changes.md
@@ -1,0 +1,1 @@
+Internal refactor to move defintion of modifierConditions closer to where they're used.
