+++ packages/adapter-mongoose/lib/tokenizers/simple.js
@@ -1,10 +1,6 @@
-const { objMerge } = require('@keystone-alpha/utils');
+const { objMerge, getType, escapeRegExp } = require('@keystone-alpha/utils');
 
-const simpleTokenizer = ({ getRelatedListAdapterFromQueryPath, modifierConditions }) => (
-  query,
-  queryKey,
-  path
-) => {
+const simpleTokenizer = ({ getRelatedListAdapterFromQueryPath }) => (query, queryKey, path) => {
   // NOTE: We slice the last path segment off because we're interested in the
   // related list, not the field on the related list. ie, if the path is
   // ['posts', 'comments', 'author', 'name'],
@@ -31,6 +27,52 @@ const simpleTokenizer = ({ getRelatedListAdapterFromQueryPath, modifierCondition
   // Nothing found, return an empty operation
   // TODO: warn?
   return {};
+};
+
+const modifierConditions = {
+  // TODO: Implement configurable search fields for lists
+  $search: value => {
+    if (!value || (getType(value) === 'String' && !value.trim())) {
+      return undefined;
+    }
+    return {
+      $match: {
+        name: new RegExp(`${escapeRegExp(value)}`, 'i'),
+      },
+    };
+  },
+
+  $orderBy: (value, _, listAdapter) => {
+    const [orderField, orderDirection] = value.split('_');
+
+    const mongoField = listAdapter.graphQlQueryPathToMongoField(orderField);
+
+    return {
+      $sort: {
+        [mongoField]: orderDirection === 'DESC' ? -1 : 1,
+      },
+    };
+  },
+
+  $skip: value => {
+    if (value < Infinity && value > 0) {
+      return {
+        $skip: value,
+      };
+    }
+  },
+
+  $first: value => {
+    if (value < Infinity && value > 0) {
+      return {
+        $limit: value,
+      };
+    }
+  },
+
+  $count: value => ({
+    $count: value,
+  }),
 };
 
 module.exports = { simpleTokenizer };
